+++ moose/include/arch/amd64/ata.h
@@ -10,4 +10,4 @@ enum ata_dev_kind {
 };
 
 #define ATA_PIO_BLKSZ 512
-int ata_pio_read(void *buf, size_t lba, size_t sector_count);
+int ata_pio_read(void *buf, u32 lba, u8 sector_count);
